pyproject.toml
# pyproject.toml
[project]
name = "clinicalmind"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.29",
  "langgraph>=0.2",
  "langchain-openai>=0.1",
  "langchain-core>=0.2",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pgvector>=0.3",
  "redis[hiredis]>=5.0",
  "tiktoken>=0.7",
  "pandas>=2.2",
  "pydantic-settings>=2.2",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "boto3>=1.34",
  "pypdf>=4.0",
  "jinja2>=3.1",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "httpx>=0.27"]